% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{GetCoverage}
\alias{GetCoverage}
\title{GetCoverage
To compute the coverage of module or pathway in the current ko profile}
\usage{
GetCoverage(pro, con, database)
}
\arguments{
\item{pro, }{ko profile}

\item{con, }{group information}

\item{database}{, database path}
}
\value{

}
\description{
GetCoverage
To compute the coverage of module or pathway in the current ko profile
}
