% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{GetReporterScore}
\alias{GetReporterScore}
\title{GetReporterScore
Computer the report score}
\usage{
GetReporterScore(dat, perm, cov, occ = 0.4)
}
\arguments{
\item{occ, }{cutoff of coverage on pathway or module}
}
\value{

}
\description{
GetReporterScore
Computer the report score
}
