% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{Getzscore}
\alias{Getzscore}
\title{GetZscore
to compute the zscore of wilcoxon test}
\usage{
Getzscore(dat, con, paired = F, hist = F)
}
\arguments{
\item{paired}{}
}
\value{

}
\description{
GetZscore
to compute the zscore of wilcoxon test
}
