% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{PermZscore}
\alias{PermZscore}
\title{PermZscore
permutation of szscore value}
\usage{
PermZscore(dat, cov, perm = 1000)
}
\arguments{
\item{dat, }{zscore result}

\item{perm, }{permuation number}
}
\value{

}
\description{
PermZscore
permutation of szscore value
}
\examples{

}
