% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{ReporterScore}
\alias{ReporterScore}
\title{ReporterScore
The command to compute the reportscore}
\usage{
ReporterScore(pr, grp, paired = F, database = "./database", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
ReporterScore
The command to compute the reportscore
}
