% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{WilcoxTest}
\alias{WilcoxTest}
\title{WilcoxTest
wilcox test or pair wilcox test}
\usage{
WilcoxTest(data, grp, paired = FALSE)
}
\arguments{
\item{data, }{{data.frame or matrix}, ko profile: row is KO ID, col is sample ID}

\item{grp, }{{data.frame}, group information, two class}

\item{paired}{, True or FALSE}
}
\value{
Wilcoxon test result
}
\description{
WilcoxTest
wilcox test or pair wilcox test
}
